+++ app/models/twitter_db/friendship.rb
@@ -4,8 +4,9 @@ module TwitterDB
     belongs_to :friend, primary_key: :uid, foreign_key: :friend_uid, class_name: 'TwitterDB::User'
 
     def self.import_from!(twitter_user)
-      friendships = twitter_user.friends.pluck(:uid).map.with_index { |uid, i| [uid, twitter_user.uid, i] }
       user = TwitterDB::User.find_or_import_by(twitter_user)
+      friendships = twitter_user.friends.pluck(:uid).map.with_index { |uid, i| [uid, twitter_user.uid, i] }
+      return if friendships.empty?
 
       ActiveRecord::Base.transaction do
         delete_all(user_uid: twitter_user.uid)
